CVSL-1906 move endpoint to existing caseload controller
